+++ src/base_formatter/builders.rs
@@ -3,9 +3,9 @@ use crate::base_formatter::format_element::tag::{
 };
 use crate::base_formatter::format_element::*;
 use crate::base_formatter::{
-    Argument, Arguments, Format, FormatResult, Formatter, GroupId, format_element,
+    Argument, Arguments, Buffer, BufferExtensions, Format, FormatResult, Formatter, GroupId,
+    VecBuffer, format_element,
 };
-use crate::base_formatter::{Buffer, BufferExtensions, VecBuffer};
 use crate::write;
 use std::cell::Cell;
 use std::marker::PhantomData;
